token auth added to routes

+++ src/controllers/categories.ts
@@ -14,18 +14,6 @@ export const getCategories = (req: Request, res: Response, next: NextFunction) =
     })
 }
 
-export const getUserCategories = (req: Request, res: Response, next: NextFunction) => {
-    const userId = req.params.userId
-    UserCategory.findAll({ where: { userId: userId }})
-    .then(result => {
-        res.json(result)
-        console.log(result)
-    })
-    .catch(err => {
-        console.log(err)
-    })
-}
-
 export const createCategory = (req: Request, res: Response, next: NextFunction) => {
     const name = req.body.name
 
@@ -41,9 +29,27 @@ export const createCategory = (req: Request, res: Response, next: NextFunction) 
     })
 }
 
+export const getUserCategories = (req: Request, res: Response, next: NextFunction) => {
+    const userId = req.params.userId
+    if (userId !== req.userId) {
+        return
+    }
+    UserCategory.findAll({ where: { userId: userId }})
+    .then(result => {
+        res.json(result)
+        console.log(result)
+    })
+    .catch(err => {
+        console.log(err)
+    })
+}
+
 export const createUserCategory = (req: Request, res: Response, next: NextFunction) => {
     const userId = req.params.userId
     const name =  req.body.name
+    if (userId !== req.userId) {
+        return
+    }
     UserCategory.create({
         name: name,
         userId: userId
@@ -59,6 +65,10 @@ export const createUserCategory = (req: Request, res: Response, next: NextFuncti
 export const editUserCategory = (req: Request, res: Response, next: NextFunction) => {
     const categoryId = parseInt(req.params.categoryId)
     const newName = req.body.newCategoryName
+    const userId = req.query.userId
+    if (userId !== req.userId) {
+        return
+    }
     UserCategory.update({ name: newName }, {where: { id: categoryId }})
     .then(newCategory => {
         res.json(newCategory)
@@ -70,6 +80,10 @@ export const editUserCategory = (req: Request, res: Response, next: NextFunction
 
 export const deleteUserCategory = (req: Request, res: Response, next: NextFunction) => {
     const categoryId = req.params.categoryId
+    const userId = req.query.userId
+    if (userId !== req.userId) {
+        return
+    }
     UserCategory.destroy({ where: { id: categoryId }})
     .then(result => {
         res.json(result)
